+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,9 +1,10 @@
 # 2.1.8
-###### November XX, 2021
+###### December XX, 2021
 - Added a `LDtkPostProcessor` system for custom control over the LDtk import process via scripting
 - Added a new toggle to the importer inspector to choose whether or not to build a color background
 - Fixed an issue where the importer inspector's section dropdowns would not remember if they were toggled
 - Added a new property for `LDtkComponentLevel` to get the level's rectangle area
+- Updated the documentation accordingly with the new features
 
 # 2.1.7
 ###### October 31, 2021 🎃